refactor: 커맨드 실행을 Main 에서 하도록 수정
