--- src/main/java/chess/CommandRun.java
@@ -1,10 +1,0 @@
-package chess;
-
-import chess.controller.ChessController;
-
-@FunctionalInterface
-public interface CommandRun<T, U> {
-
-    void execute(T t, U u);
-
-}
